Visitor announcement and panic button.
The visitor announcement has been implemented and is working correctly in the both apartments.

Also, the panic button is working in the both apartments, but it remains to implement some way to send an e-mail.

Each apartment, and the gatehouse, has its own GUI and all is prepared to run in two different computers.